backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
kombu==5.5.4
oauthlib==3.3.1
openai==1.93.0
packaging==25.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
